+++ packages/markdown/src/hooks/useContent.js
@@ -1,8 +1,0 @@
-import useExternalContent from './useExternalContent';
-
-export default function useContent(children, src, loadExternal) {
-    if (src) {
-        return useExternalContent(src, loadExternal);
-    }
-    return children;
-}