+++ packages/markdown/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.2.0
+
+- Drop local content loading code, use `@loopmode/use-content` instead
+- Rename `loadExternal` prop to `loader`
+
 ### 1.1.0
 
 - Add `boxed` prop